完成了服务端的重构工作，主要进行了以下几部分的重构： 首先是将业务代码进行包装，对外界只可以通过map表查找对应的业务逻辑块 注意到每次都必须手动的去调用单例发送内容这一过程相对繁琐，所以通过链表存储需要发送的包，将所有的包放到链表中，然后在统一释放，这一过程的实现主要是通过回调函数实现的，通过该函数可以完成抓哟的业务逻辑内容，且将包放入到链表中，最终做统一的释放。针对需要使用的包，采用末尾参数的灵活使用实现
